+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/movimentacao/_model/OrigemDestinoEnum.java
@@ -1,25 +1,32 @@
- package br.com.zalf.prolog.webservice.frota.pneu.movimentacao._model;
+package br.com.zalf.prolog.webservice.frota.pneu.movimentacao._model;
 
 import br.com.zalf.prolog.webservice.frota.pneu._model.StatusPneu;
+import br.com.zalf.prolog.webservice.frota.pneu.movimentacao.motivos.transicao._model.listagem.TransicaoUnidadeMotivos;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
- /**
- * Created by Zart on 02/03/17.
+/**
+ * Enum que representa as origens e destinos possíveis nas movimentações do Prolog.
+ * Para facilitar a passagem dessas constantes para a tabela PNEU, usamos como valor do enum as constantes do
+ * enum {@link StatusPneu}.
+ *
+ * <b>A ordem das constantes nesse enum é importante, não altere!</b>
+ * Ele é utilizado para ordenação de objetos no método:
+ * {@link TransicaoUnidadeMotivos#compareTo(TransicaoUnidadeMotivos)}
  */
 public enum OrigemDestinoEnum implements PermissoesValidatorString {
-    VEICULO(StatusPneu.EM_USO) {
-        @NotNull
-        @Override
-        public String getStringPermissoesValidator() {
-            return "VEÍCULO";
-        }
-    },
     ESTOQUE(StatusPneu.ESTOQUE) {
         @NotNull
         @Override
         public String getStringPermissoesValidator() {
             return "ESTOQUE";
+        }
+    },
+    VEICULO(StatusPneu.EM_USO) {
+        @NotNull
+        @Override
+        public String getStringPermissoesValidator() {
+            return "VEÍCULO";
         }
     },
     DESCARTE(StatusPneu.DESCARTE) {
@@ -44,17 +51,17 @@ public enum OrigemDestinoEnum implements PermissoesValidatorString {
         this.statusPneu = statusPneu;
     }
 
-     @NotNull
-     public static OrigemDestinoEnum fromString(@Nullable final String text) throws IllegalArgumentException {
-         if (text != null) {
-             for (final OrigemDestinoEnum origemDestinoEnum : OrigemDestinoEnum.values()) {
-                 if (text.equalsIgnoreCase(origemDestinoEnum.statusPneu.asString())) {
-                     return origemDestinoEnum;
-                 }
-             }
-         }
-         throw new IllegalArgumentException("Nenhuma origem/destino encontrada para a String: " + text);
-     }
+    @NotNull
+    public static OrigemDestinoEnum fromString(@Nullable final String text) throws IllegalArgumentException {
+        if (text != null) {
+            for (final OrigemDestinoEnum origemDestinoEnum : OrigemDestinoEnum.values()) {
+                if (text.equalsIgnoreCase(origemDestinoEnum.statusPneu.asString())) {
+                    return origemDestinoEnum;
+                }
+            }
+        }
+        throw new IllegalArgumentException("Nenhuma origem/destino encontrada para a String: " + text);
+    }
 
     @NotNull
     public String asString() {